feat : api 연동하고 , type 지정
